backend-django/requirements.txt
asgiref==3.11.1
Django==6.0.4
django-cors-headers==4.9.0
django-restframework==0.0.1
djangorestframework==3.17.1
sqlparse==0.5.5
tzdata==2026.2
